fix: replace fs.watch with chokidar
fs.watch has presented numerous compatibility issues that cause failures
in some environments and machines. This seems to most prominently
affect Linux distros but not providing Node with the adaquate means of
recursively watching directories. As a result we need to replace the
watch features with something more compatible. Chokidar was chosen as
it has a proven record since beeing installed in numerous JS based
systems, such as VSCode and webpac