fix: create a multisite htaccess file when enabled
htaccess files require different configuration depending on whether
multisite is enabled and based on which type - subfolder or subdomain.
Previously only a single site htaccess file was created no matter the
config. To correct this I've introduced two multisite .htaccess
templates that will be used when their type is selected.

* Added subdomain .htaccess
* Added subfolder htaccess
* Change dockerfile tempalte to use htaccess name variable
* Change config creation to pass htaccess file name to template